3rd step for git retries
- instanciate in application with hardcoded retry policy

+++ cf-ops-automation-broker-core/src/main/java/com/orange/oss/cloudfoundry/broker/opsautomation/ondemandbroker/git/RetrierGitManager.java
@@ -37,7 +37,9 @@ public class RetrierGitManager implements GitManager {
                 gitManager.commitPushRepo(ctx, false);
             });
         } finally {
-            gitManager.deleteWorkingDir(ctx);
+            if (deleteRepo) {
+                gitManager.deleteWorkingDir(ctx);
+            }
         }
     }
